tear down & structure framebuffer render process, framebuffer works, some problems,
