+++ src/game/World.h
@@ -9,7 +9,7 @@
 #include "ECS.h"
 
 
-#define WORLD_DIAMETER (32)
+#define WORLD_DIAMETER (16)
 struct World {
     explicit World(std::unique_ptr<WorldSave> sv);
