+++ src/main.cpp
@@ -6,6 +6,9 @@
 #include "rendering/RegionRenderer.h"
 #include "rendering/WorldRenderer.h"
 #include "game/World.h"
+#include "render_pipeline_structs.h"
+#include "rendering/BufferRenderer.h"
+#include "rendering/PostRenderer.h"
 LOG_MODULE(main);
 
 #include <iostream>
@@ -23,10 +26,6 @@ public:
     ~Scopetimer() {float time = sw.stop(); LOG_DBG("%s time %.0fus", name.c_str(), time);}
 };
 
-
-vec2 wcampos {0.f,0.f};
-float wcamvw = REGION_SIZE*2;
-
 class WorldDriver : public GameDriver {
 public:
 	WorldDriver();
@@ -39,57 +38,58 @@ private:
 
 	World world;
 
-	WorldRenderer wrenderer;
+	BufferRenderer brenderer;
+	PostRenderer prenderer;
 	TextRenderer text_renderer;
 
-	OrthoCamera& cam;
+	local_cam_t lcam;
 
 };
 
 WorldDriver::WorldDriver() : GameDriver(),
-							world(std::make_unique<MapWSave>(std::make_unique<TestWorldGenerator>(0xFACEFACE))),
-							cam(wrenderer.cam)
+							world(std::make_unique<MapWSave>(std::make_unique<TestWorldGenerator>(0xFACEFACE)))
 							{
 							}
 
 void WorldDriver::user_create() {
 	Renderer::context_init("untitled", 720, 720);
-	cam.update();
-	wrenderer.use_world(world);
-	wrenderer.init();
+	lcam.frame = vec2(REGION_SIZE*2.f, (REGION_SIZE*2.f)/window.aspect);
+	(*(brenderer.input_ptr())) = {
+		.world = &world,
+		.wmpos = vec2(0.f) /* TODO fuck */,
+		.lcam = this->lcam
+	};
+	brenderer.init();
+	prenderer.init();
+	LOG_DBG("renderers init");
 	text_renderer.init_text_rendering();
 	text_renderer.init();
 	text_renderer.set_text("move with wasd | scroll to zoom | click to place walls\npress L to turn on lighting (bad) (wip) %s", "");
+	LOG_DBG("user driver created");
 }
 
 void WorldDriver::user_update(float dt, Keyboard const& kb, Mouse const& mouse) {
 	if (kb[GLFW_KEY_ESCAPE].down) this->close();
-	if (kb[GLFW_KEY_W].down) wcampos.y += dt * (.1f + ((kb[GLFW_KEY_LEFT_SHIFT].down) * 32.f));
-	if (kb[GLFW_KEY_A].down) wcampos.x -= dt * (.1f + ((kb[GLFW_KEY_LEFT_SHIFT].down) * 32.f));
-	if (kb[GLFW_KEY_S].down) wcampos.y -= dt * (.1f + ((kb[GLFW_KEY_LEFT_SHIFT].down) * 32.f));
-	if (kb[GLFW_KEY_D].down) wcampos.x += dt * (.1f + ((kb[GLFW_KEY_LEFT_SHIFT].down) * 32.f));
-	world.relocate(wcampos);
+	if (kb[GLFW_KEY_W].down) lcam.pos.y += dt * (.1f + ((kb[GLFW_KEY_LEFT_SHIFT].down) * 32.f));
+	if (kb[GLFW_KEY_A].down) lcam.pos.x -= dt * (.1f + ((kb[GLFW_KEY_LEFT_SHIFT].down) * 32.f));
+	if (kb[GLFW_KEY_S].down) lcam.pos.y -= dt * (.1f + ((kb[GLFW_KEY_LEFT_SHIFT].down) * 32.f));
+	if (kb[GLFW_KEY_D].down) lcam.pos.x += dt * (.1f + ((kb[GLFW_KEY_LEFT_SHIFT].down) * 32.f));
+	world.relocate(lcam.pos);
 
-	if (kb[GLFW_KEY_K].pressed) {static bool wf = 1;wrenderer.twf();wf = !wf;}
 	if (abs(mouse.scroll.y) > 0.1) {
 #ifdef __APPLE__
 		float vwadd = dt * mouse.scroll.y * 100.f;
 #else
 		float vwadd = dt * mouse.scroll.y * 1000.f;
 #endif
-		// uint32_t oldpw = wrenderer.get_pix_width();
-		// wrenderer.set_pix_width(
-		// 		glm::clamp((int32_t)wrenderer.get_pix_width() + ((int32_t)(vwadd)), 2, 5*REGION_SIZE*16*(WORLD_DIAMETER-1))
-		// );
-		wcamvw += vwadd;
+		local_cam_setvw(lcam, glm::clamp(lcam.frame.x + vwadd, 8.f, REGION_SIZE*512.f));
+		// LOG_INF("changing vw to %.1f,%.1f", lcam.frame.x, lcam.frame.y);
 	}
 
-	// if (kb[GLFW_KEY_DOWN].pressed) wrenderer.set_pix_width(wrenderer.get_pix_width()-1);
 
-	cam.update();
-
+	// ugh TODO fix mouse
 	if (mouse.left.down) {
-		vec2 ssm = world.world_mpos(mouse.pos, window.frame, &cam);
+		vec2 ssm = world.world_mpos(mouse.pos, window.frame, &brenderer.cam);
 		Tile& tile = world.tile_at(vec2(ssm.x, ssm.y));
 		tile.surf = {
 			.img = 3,
@@ -107,13 +107,25 @@ void WorldDriver::user_update(float dt, Keyboard const& kb, Mouse const& mouse) 
 
 	static size_t fpt = 0;
 	fpt++;
-	if (!(fpt % 60)) LOG_INF("fps: %.1f", 1./dt);
+	if (!(fpt % 1024)) LOG_INF("fps: %.1f", 1./dt);
 }
 
 void WorldDriver::user_render() {
-	wrenderer.give_mouse(window.mouse.pos);
-	wrenderer.prepare();
-	wrenderer.render();
+	// input data from engine into buf renderer
+	(*(brenderer.input_ptr())) = {
+		.world = &world,
+		.wmpos = vec2(0.f) /* TODO fuck */,
+		.lcam = this->lcam
+	};
+	// buffer render to framebuffer
+	brenderer.prepare();
+	brenderer.render();
+	// transfer buf render output to post render input
+	brenderer.write_output(prenderer.input_ptr());
+	// world frame to window
+	prenderer.prepare();
+	prenderer.render();
+
 	text_renderer.render(24, window.frame.y - 48, 4);
 }
